refactor(constants): reclassify protocol constants (ADR-043) (#1360)
* refactor(constants): reclassify protocol constants per ADR-043

Phase 3 (Rust core):
- Remove MAX_NESTING_DEPTH protocol ceiling — unbounded by default,
  context-configurable via ContextParams::max_nesting_depth
- Remove PROTOCOL_HARD_MAX_CHAIN_DEPTH — chain depth is context-
  configurable, DEFAULT_MAX_CHAIN_DEPTH raised 3→8
- DEFAULT_SESSION_CAP_PER_CALLER raised 5→1000, type usize→u32,
  context-configurable via ContextParams::session_cap
- RepublishConfig: relay_blob_ttl_secs configurable, republish
  interval derived as max(ttl-86400, ttl/2, 60)
- SyncPolicy constants demoted to implementation recommendations

Phase 4 (FFI + SDKs):
- WASM: remove PROTOCOL_HARD_MAX_CHAIN_DEPTH, update defaults
- All bridges: pass through max_nesting_depth and session_cap
- TypeScript: widen chainDepth validation 0-5 → 0-255
- Kotlin: update doc comments on sync constants

Phase 5 (Tests):
- Update all assertions for new defaults and removed ceilings
- Add tests for context-configurable nesting/session/chain depth
- Add republish interval derivation tests including small TTLs
- Fix ~20 stale comments with old constant values


* docs(specs): reclassify protocol constants (ADR-043)

ADR-043: Reclassify 11 protocol constants into three tiers:
- Protocol Invariants: DHT republish, sender key grace, bucket sizes
- Configurable Parameters: nesting depth (unbounded), chain depth
  (default 8, no hard max), session cap (default 1000), relay TTL
- Implementation Recommendations: sync timeouts, catch-up limit

Remove PROTOCOL_HARD_MAX_CHAIN_DEPTH and MAX_NESTING_DEPTH as protocol
ceilings. Both are now context-configurable with no protocol ceiling.
DEFAULT_MAX_CHAIN_DEPTH raised from 3 to 8.

Spec sections updated: §9.18 (registry restructured), §9.2.1, §5.13.8,
§6.2.1, §24.1, §24.4, §23.


* docs(specs): reclassify protocol constants (ADR-043)

ADR-043: Reclassify 11 protocol constants into three tiers:
- Protocol Invariants: DHT republish, sender key grace, bucket sizes
- Configurable Parameters: nesting depth (unbounded), chain depth
  (default 8, no hard max), session cap (default 1000), relay TTL
- Implementation Recommendations: sync timeouts, catch-up limit

Remove PROTOCOL_HARD_MAX_CHAIN_DEPTH and MAX_NESTING_DEPTH as protocol
ceilings. Both are now context-configurable with no protocol ceiling.
DEFAULT_MAX_CHAIN_DEPTH raised from 3 to 8.

Spec sections updated: §9.18 (registry restructured), §9.2.1, §5.13.8,
§6.2.1, §24.1, §24.4, §23.


---------

+++ crates/scp-core/src/context/tools/mod.rs
@@ -267,7 +267,7 @@ pub enum ToolError {
         max_depth: u8,
     },
 
-    /// The per-caller session cap has been reached (spec section 6.2.1, 9.2.1).
+    /// The per-caller session cap has been reached (spec §6.2.1, §9.2.1, ADR-043).
     #[error(
         "session cap exceeded: calling context \"{source_context}\" has {current} active sessions (max {max})"
     )]
@@ -275,9 +275,9 @@ pub enum ToolError {
         /// The calling context that hit the cap.
         source_context: String,
         /// Current number of active sessions from this caller.
-        current: usize,
+        current: u32,
         /// Maximum allowed concurrent sessions per caller.
-        max: usize,
+        max: u32,
     },
 
     /// The tool schema does not meet the specificity floor (spec section 6.2, 9.2.1).
